rename kvpair as mongomerkle
1. we will like to store data-hash pair separately
2. we will save hash in mongomerkle and use a separate dbcollection for
   kv: hash -> data
